functions/most_popular_repo_strands/requirements.txt
strands-agents==1.24.0
requests==2.32.5
PyGithub==2.8.1
boto3==1.42.39
beautifulsoup4==4.14.3
